requirements.txt
requests           2.31.0
mysql-connector    2.2.9